Recipes

Recipes Docs

MicrobiomePlots.AnnotationBar — Type.

AnnotationBar(::Array{<:Color,1})

Array of colors for plotting a grid. Use plot(::AnnotationBar; kwargs...)

source

MicrobiomePlots.annotationbar — Method.

annotationbar(colors::AbstractArray{<:Color,1})
annotationbar(labels::AbstractArray{<:AbstractString,1})
annotationbar(labels::AbstractArray{<:AbstractString,1}, colormap::Dict{<:AbstractString,<:Color})
annotationbar(labels::AbstractArray{<:AbstractString,1}, colors::AbstractArray{<:Color,1})

Create an AnnotationBar from a vector of colors or a vector of labels (Strings).

A vector of Colors will be used directly to create an AnnotationBar

For an array of Strings, each unique value will be assigned a random color from Colors.color_names.

A vector of labels may be passed with a Dict with keys for each label mapping to a Color value.

A vector of labels and colors may be used.

If the lengths of these vectors are the same, the colors will be used to create an AnnotationBar

If the length of labels is longer than the length of colors, each unique label will be assigned to one of the colors

source